Move capabilities to platform
