+++ src/vs/platform/terminal/common/capabilities/capabilities.ts
