+++ src/vs/workbench/contrib/terminal/test/browser/capabilities/commandDetectionCapability.test.ts
@@ -6,9 +6,9 @@
 import { deepStrictEqual, ok } from 'assert';
 import { timeout } from 'vs/base/common/async';
 import { Terminal } from 'xterm';
-import { CommandDetectionCapability } from 'vs/workbench/contrib/terminal/common/capabilities/commandDetectionCapability';
+import { CommandDetectionCapability } from 'vs/platform/terminal/common/capabilities/commandDetectionCapability';
 import { NullLogService } from 'vs/platform/log/common/log';
-import { ITerminalCommand } from 'vs/workbench/contrib/terminal/common/capabilities/capabilities';
+import { ITerminalCommand } from 'vs/platform/terminal/common/capabilities/capabilities';
 
 async function writeP(terminal: Terminal, data: string): Promise<void> {
 	return new Promise<void>((resolve, reject) => {
